pyproject.toml
[project]
name = "yuxi-know"
version = "0.2.0.dev"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "asyncpg>=0.30.0",
    "chromadb>=1.0.15",
    "colorlog>=6.9.0",
    "dashscope>=1.23.2",
    "docx2txt>=0.9",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.22",
    "langchain-deepseek>=0.1.3",
    "langchain-huggingface>=0.2.0",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.14",
    "langchain-tavily>=0.2.3",
    "langchain-text-splitters>=0.3.9",
    "langchain-together>=0.3.0",
    "langgraph>=0.3.34",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "langgraph-cli[inmem]>=0.1.54",
    "langsmith>=0.3.37",
    "lightrag-hku>=1.3.9",
    "llama-index>=0.12.33",
    "llama-index-readers-file>=0.4.7",
    "loguru>=0.7.0",
    "markdownify>=1.1.0",
    "mcp>=1.12.0",
    "neo4j>=5.28.1",
    "networkx>=3.5",
    "openai>=1.76.0",
    "opencv-python-headless>=4.11.0.86",
    # "paddleocr>=2.10.0",  # 使用远程 API 调用,不需要本地安装
    "pyjwt>=2.8.0",
    "pymilvus>=2.5.8",
    "pymupdf>=1.25.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "rapidocr-onnxruntime>=1.4.4",
    "tavily-python>=0.7.0",
    "unstructured>=0.17.2",
    "uvicorn[standard]>=0.34.2",
    "httpx>=0.27.0",
    "tqdm>=4.66.4",
    "rich>=13.7.1",
    "typer>=0.16.0",
    # "mineru[core]>=2.1.6",  # 使用远程 MinerU API 调用,不需要本地安装 torch/cuda
    "tabulate>=0.9.0",
    "minio>=7.2.7",
    "Pillow>=10.5.0",
    "pymysql>=1.1.0",
    "tenacity>=8.0.0",
    "pypinyin>=0.55.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
]
[tool.ruff]
line-length = 120  # 代码最大行宽
lint.select = [         # 选择的规则
    "F",
    "E",
    "W",
    "UP",
]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["test"]
markers = [
    "auth: marks tests that require authentication",
    "slow: marks tests as slow",
    "integration: marks tests as integration tests"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"


[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "pytest-cov>=6.0.0",
]
